[libc-trusty] Workaround a codegen bug in Clang
Certain version of Clang may emit a reference to __multi3 when compiling
__builtin_mul_overflow(uint32_t, uint32_t, &uint64_t). compiler_rt for
32-bit ARM does not contain this function. Work around the problem by
casting one of the arguments to a 64-bit integer.

Bug: 145830721

+++ lib/libc-trusty/time.c
@@ -48,7 +48,12 @@ int clock_nanosleep(clockid_t clock_id,
     }
     /* Convert timespec to nanoseconds. */
     uint64_t sleep_time;
-    if (__builtin_mul_overflow(req->tv_sec, NS_PER_SEC, &sleep_time)) {
+    /*
+     * Note: casting NS_PER_SEC to work around a Clang codegen bug.
+     * See: b/145830721
+     */
+    if (__builtin_mul_overflow(req->tv_sec, (uint64_t)NS_PER_SEC,
+                               &sleep_time)) {
         errno = EINVAL;
         return -1;
     }
